made type safe PropertyChangedListener and abstracted setters from implemetation of AccountPropertiesManager
