+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/auth/DownloadCustomCertDialogFragment.java
@@ -23,7 +23,7 @@ import org.androidannotations.annotations.Bean;
 import org.androidannotations.annotations.EFragment;
 import org.androidannotations.annotations.InstanceState;
 import org.ovirt.mobile.movirt.R;
-import org.ovirt.mobile.movirt.auth.properties.AccountPropertiesManager;
+import org.ovirt.mobile.movirt.auth.properties.manager.AccountPropertiesManager;
 import org.ovirt.mobile.movirt.ui.UiUtils;
 import org.ovirt.mobile.movirt.ui.dialogs.DialogListener;
 import org.ovirt.mobile.movirt.util.URIUtils;
